Namespacify classes
Make it so code makes use of a PHP 5.3 feature, Namespaces. I need this
because I work on a WordPress site that makes use of the LiveFyre PHP
Utils library, which in turn makes use of this Requests library. There
is a class name collision with the native WordPress Requests class,
which causes fatal errors for some operations.