Unit tests: attempt to make the build more stable
The `testAlternatePort` tests for both Curl as well as fsockopen often time out and cause builds to fail.

This PR attempts to make the builds more stable, by:
1. Retrying the request a second time if the first request timed out.
2. Skipping the unit test if the second request timed out as well.

Adding `verbose=true` to the PHPUnit config will make sure that if the tests are skipped, the skip reason will always be displayed when running the tests.