:wrench: QA: lint PHP files in this repo
This PR adds linting of the PHP files against every PHP version to the Travis build.

The `vendor` directory is excluded by default.
For PHP 5.2, the `tests/phpunit6-compat.php` file is excluded as it uses namespaces.
For PHP < 5.4, the `examples/cookie_jar.php` file is excluded as it uses short arrays.