QA: remove useless variable assignments [2]
Each of these is testing that the function call triggers an exception - which stops the test dead - or is checking an assertion via a hooked-in method.
So no need to assign the return of the function call to a variable, when the variable is not being tested via an assertion.