Tests: verify exception messages
Don't just verify that an exception is thrown, but verify that it is the exception expected for a particular test.

Note: in some cases, the verification is done on a _partial_ message string to allow for changed exception message strings across PHP versions.

In a future commit in which the annotations for exception expectations are replaced by method calls, logic can be added to improve the message checking in a cross-version compatible manner.