Requests::parse_response(): ensure URL is always a string
This fixes an issue on PHP 5.2/5.3 where the `Requests_Response::__toString()` method would not always automagically be invoked.

I've verified that this is the only place in the application where the `Requests_Response::$url` property is being set.

I considered making the property `private` and adding magic setter/getter methods, but that would open the class up to injection of other non-public properties, which is not the intention of this change.

Includes:
* Making the tests which cover this property use a strict typed assertion.
* Adding "failure messages" to the various assertions in this test to more easily be able to distinguish which assertion has failed.