PHPCS ruleset: exclude various other rules
`WordPress.Files.FileName`

This repo uses PSR-0 for the class name determination, which conflicts with the WPCS rules, so ignoring the WPCS sniff.

There is currently no sniff available to enforce PSR-0.

`Squiz.ControlStructures.ControlSignature.SpaceAfterCloseBrace`

This contrast with the code style used in the code base as-is. The opposite can in the future be enforced once PHPCSExtra 1.0 has been tagged (which will be included in WPCS 3.0.0).

Ignore various WP specific sniffs.

As Requests is a stand-alone library, sniffs referring to WordPress native functions or doing things "the WordPress way" are irrelevant.