FilteredIterator: fix PHP 5.2 compatibility and stabilize the code
The `$callback` passed to the constructor of the `FilteredIterator` class may not be callable and if so, would cause a fatal error.

This puts some defensive coding in place to guard against that and adds a unit test for it as well.

This also fixes the tests failing on PHP 5.2, as when the object is unserialized on PHP 5.2, the `$callback` property will no longer be set. This looks to have been a bug in PHP 5.2 which was fixed in PHP 5.3.