README: fix code block
The homepage of the https://requests.ryanmccue.info/ website is automatically generated from the `readme` file.

The last time the site was generated was in 2016. Since then a lot has changed in GH Pages and a lot of the [dependencies](https://pages.github.com/versions/) used to generate a site have been updated to newer releases.

Most notable, the kramdown engine has moved from v1 to v2 which has implications for how code blocks are shown on the website.

For the best results, it is strongly recommended to not use "whitespace" marked code blocks, but to use code block fencing with a language indicator to enable syntax highlighting as well.

This commit fixes up the `README` file in anticipation of the release.