GH Pages: make the API documentation more discoverable
At this moment, there is only one link to the API documentation and it is hidden away on the first page of the "Documentation" page where it can be easily overlooked.

This adds a new menu item in the main menu at the top of the page linking to the API Docs for improved discoverability.