GH Pages: add Gemfile
A GH Pages site needs a Gemfile to be able to generate and to allow for testing locally.

It may be that in the past this was not a hard requirement, but either way, it is good practice to have the file in place.

Includes adding the `lock` file to `.gitignore` and both the `Gemfile` and the `lock` file to the list of files which should not be copied over into the live website.