Docs: update content and various tweaks to the markdown
The "Documentation" section of the https://requests.ryanmccue.info/ website is automatically generated from the files in the `docs` folder.

The last time the site was generated was in 2016. Since then a lot has changed in GH Pages and a lot of the [dependencies](https://pages.github.com/versions/) used to generate a site have been updated to newer releases.

This commit makes various tweaks to the markdown for improved rendering on the current GH Pages setup.
While going through the pages, I've also done a cursory review of the content and fixed up information which was extremely outdated or plain incorrect by now.

Includes:
* Use spaces not tabs for indentation in markdown files.
* Fixing up various URLs/links.
* Removal of various uses of ableist language.
* Punctuation fixes.