max_bytes not respected in a performant manner
With some testing I was noticing that a request for 100 byes from a 10MB file was timing out after 5 seconds.

Turns out that the cause was that Requests downloads the full 10MB file in order to return the first 100 bytes of it.

This PR simply aborts the connection when enough bytes have been read from the stream.

Co-authored-by: Dion Hulse <contact@dd32.id.au>