Drop support for HHVM
HHVM dropped support for PHP at the end of 2018.

While testing and basic support for HHVM was enabled in Request 1.7.0, it has basically been broken since then and we are currently no longer testing against HHVM anyway.

So, let's formally drop support an remove all references to it (aside from those in the changelog).

Ref: https://hhvm.com/blog/2017/09/18/the-future-of-hhvm.html