Deprecated classes classmap file: wrap each class/interface declaration in an "exists" check
This fixes a potential fatal "class already exists" error when during one "page load":
* Two projects both using Requests are used.
* Neither of these projects uses Requests with a custom namespace/prefix as can be done via PHP_Scoper.
* The first of these projects loaded is still using Requests 1.x and is using Composer autoload.
* The second project is using Requests 2.x and is also using Composer autoload,

In that case, depending on the order of the registered autoloaders, the `Deprecated.php` file _may_ be called while _some_ Requests 1.x have already been autoloaded, but not all.

Bulk declaring the deprecated classes will then lead to a "Fatal error: Cannot declare interface/class Requests_..., because the name is already in use" error.

Using `class_exists()` wrappers will prevent this.

In addition to this, the conditions also check for the existance of the "new" classes to prevent potential fatal errors when only a selection of the classes from this library are used/available within an external project.