FsockopenTest: add regression test for HTTP string
PR 339 made a change to safeguard that the HTTP string would always contain a decimal point, never a comma.

The original bug was related to the current locale influencing the results of the call to `sprintf()`.

At the time, no test was added to safeguard the fix against regressions.

This PR now adds this regression test.

I've verified that without the fix as merged in 339, this test would fail.

Includes adjusting the GH Actions workflows to install extra locales as otherwise the test would fail on the locale being changed to in the test not being available.