Add issue templates
Add an initial set of basic issue templates.

The initial set consists of a basic _bug report_ and a _feature request_ template, loosely based on the standards proposed by GitHub itself, with adjustments based on experiences in other repos.

The intention is to initially iterate on these templates based on how well people fill them out and how helpful they prove to be to get bugs reported in a reproducible format.
Once the templates are relatively stable, we can consider switching them over to Yaml template format.