Requests: set default certificate path as a constant
This value does not change during the request, so should be a constant and as the minimum PHP version is now PHP 5.6, we can use constant scalar expressions in constants, so don't have to worry about the concatenation.

Includes moving this static file to a separate directory, so the `src` directory only contains code, not assets, like the certificates file and updating references to the file location elsewhere in the code.