IdnaEncoder::encode(): add input validation
The `IdnaEncoder::encode()` is the official entry point for this class, even though there are more public methods (of which it is questionable whether they should be public, but changing that now would be a BC-break).

The documented accepted parameter type is `string`, but no input validation was done on the parameter, which could lead to various PHP errors, most notably a "passing null to non-nullable" deprecation notice on PHP 8.1.

This commit adds input validation to the `IdnaEncoder::encode()` method, allowing only for strings and _stringable_ objects.

While this is stricter than previously, only a string could result in a valid return value, so in my opinion, being strict is warranted.

Includes adding perfunctory tests for the new exception.