Exception\Http\StatusUnknown: minor bug fix
The default value of the `Requests\Response::$status_code` property is `false` and that value is also meant to be used for non-blocking errors.

If, in that case, as `StatusUnknown` error would be thrown, the `$error_code` for the Exception would be set to the invalid value `false`, while this should always be an integer.

Fixed now.