Hooks: add input validation to all methods
Neither of the methods in the `WpOrg\Requests\Hooks` class did any input validation. This could lead to various PHP errors and becomes all the more relevant due to the new "passing null to non-nullable" deprecation notice on PHP 8.1.

This commit adds simple input validation to all methods within the `WpOrg\Requests\Hooks` class and will throw an `WpOrg\Requests\Exception\InvalidArgument` exception if a parameter does not comply with the expectations.

Includes adding tests for the new exceptions and a test to safeguard that closures are accepted