Documentation: fix references to Curl resources
The PHP Curl extension worked with a Curl resource in PHP 4/5/7. As of PHP 8.0, `curl_init()` returns an instance of `CurlHandle` instead of a resource.

This updates the documentation to indicate that either a `resource` or `CurlHandle` instance can be expected in certain places in the code base (and that both are supported).

Ref: https://www.php.net/manual/en/migration80.incompatible.php#migration80.incompatible.resource2object