GH Pages: add "outdated" warning to the Requests 1.x docs
The "old" docs will no longer be updated (as the branch won't be updated anymore), but this way they will still be available if people are looking for them and links around the web to the "old" docs will continue to work.

This commit:
* Adds a new folder `api-2.x` for the _new_ docs.
* Adjusts the menu at the top of the page of the website to provide access to both the `1.x` as well as the `2.x` docs.
* Adds a "You are looking at the documentation for an old version of Requests" warning at the top of each page for the "old" docs.
    For all, but the `source-class-....html` files, the warning will contain a direct link to the equivalent page in the `2.x` docs.
* Adds minimal styling for the warning to the `style.css` file.

Note: the `2.x` API docs will automatically be added/updated on the release, so no files have been added to the `api-2.x` folder, other than a `.gitkeep` file to prevent having to adjust the workflow to account for the directory potentially not existing. This `.gitkeep` file will be automatically removed the first time the 2.x docs are generated and committed.