QA/PHP 8.0 | Jar::before_redirect_check(): don't use reserved keywords as param names
While using reserved keywords as parameter names is perfectly valid PHP, I would strongly recommend against it as this decreases the comprehensibility of code when these parameter names are subsequently used in function calls using named parameters (PHP 8.0+).

This changes the parameter name for the `Jar::before_redirect_check()` function from `$return` to `$response` and updates all uses of the parameter in the function to match.

Includes fixing the docblock, in which the parameter name already didn't match the previous parameter name.