QA/PHP 8.0 | Cookie::domain_matches(): don't use reserved keywords as param names
While using reserved keywords as parameter names is perfectly valid PHP, I would strongly recommend against it as this decreases the comprehensibility of code when these parameter names are subsequently used in function calls using named parameters (PHP 8.0+).

This changes the parameter name for the `Cookie::domain_matches()` function from `$string` to `$domain` and updates all uses of the parameter in the function to match.

Includes changing the name of the function local `$domain_string` variable to `$cookie_domain` and updating the inline comments for comprehensibility as they had become unclear with the new names in play.