QA/PHP 8.0 | Cookie::parse(): don't use reserved keywords as param names
While using reserved keywords as parameter names is perfectly valid PHP, I would strongly recommend against it as this decreases the comprehensibility of code when these parameter names are subsequently used in function calls using named parameters (PHP 8.0+).

This changes the parameter name for the `Cookie::parse()` function from `$string` to `$cookie_header` and updates all uses of the parameter in the function to match.