QA/PHP 8.0 | IdnaEncoder::nameprep(): don't use reserved keywords as param names
While using reserved keywords as parameter names is perfectly valid PHP, I would strongly recommend against it as this decreases the comprehensibility of code when these parameter names are subsequently used in function calls using named parameters (PHP 8.0+).

This changes the parameter name for the `IdnaEncoder::nameprep()` function from `$string` to `$text` and updates all uses of the parameter in the function to match.