QA/PHP 8.0 | Test bootstrap: don't use reserved keywords as param names
While using reserved keywords as parameter names is perfectly valid PHP, I would strongly recommend against it as this decreases the comprehensibility of code when these parameter names are subsequently used in function calls using named parameters (PHP 8.0+).

While this use in the test bootstrap is not problematic, for consistency, it is changed as well.