EncodingTest: move to subdirectory and rename
There are a number of test classes which all directly relate to the `WpOrg\Requests\Requests` class.

Having those test files/classes all in the `tests` root directory does not make it any clearer what these tests are actually testing. So, I'm proposing that if a class warrants multiple test classes, we create a directory named after the class and place the files in that directory.

Also, the tests in this file are not about _encoded_ data, but about _compressed_ data, so renaming the test class to be more descriptive.