Requests::decompress()/compatible_gzinflate(): add input validation
The documented accepted parameter type for both these methods is `string`, but no input validation was done on the parameter, which could lead to various PHP errors, most notably a "passing null to non-nullable" deprecation notice on PHP 8.1.

This commit adds input validation to the `Requests::decompress()` and the `Requests::compatible_gzinflate()` methods, allowing only for strings.
Note: Stringable objects will not be considered as valid input for these methods.

Includes adding perfunctory tests for the new exceptions.