GH Actions: more website update workflow tweaks
1. Change the workflow trigger

As it was, the website update PR would be created when a _release_ was published, while from a "human workflow" point of view, it could be considered better to create the website update PR when a new tag is pushed.
That way, the website update PR can be reviewed and merged before the release notification is created and mailed out to subscribers/watchers of the repo.

2. Make messages more descriptive

The `github.ref` variable contains the tag on release, the branch name on pushes and the PR merge branch name of pull requests. All of these generally prefixed with `refs/pull/`, `refs/heads/` etc
Those prefixes decrease the readability of the messages/branch names in which the variable is used.

By using `github.ref_name` for pulls and `github.event.release.tag_name` for tags, we should get the "clean" branch/tag name instead.