BC-layer: fix fatal "can not inherit from final class" errors
In PR #534 (to address #514), a number of classes were made `final`. This effectively broke the BC-layer for the Composer autoloader as that BC-layer relied on the "old" class names extending the new ones.

Fixed now by:
* Using the new custom autoloader - which uses class aliasing, which is not affected by the change to `final` - for the BC-layer for Composer.
* Automatically loading the file via the composer `autoload` directive.

While using the `files` directive in Composer `autoload` is generally a last resort, as this file only registers an autoloader and Composer autoload is used for just that purpose, this should not be a problem.