Autoload: treat PSR-4 case-sensitively and PSR-0 case-insensitively
As the most common server OS-es have a case-sensitive file system and PSR-4 class loading does a direct translation from the class name to the file name, the PSR-4 based autoloading is now completely case-sensitive.

As for the BC-layer for PSR-0: this will now consistently be handled case-insensitively.
This was largely so, what with the case-insensitive classmap being used. It may also make the change over slightly easier as a number of class name have changes in the case being used.

Includes a minor adjustment to an existing test to safeguard the case-insensitive handling handling of the PSR-0 classes.