Class Ipv6: add input validation
The documented accepted parameter type for all methods in the `Ipv6` class is `string`, but no input validation was done on the parameter, which could lead to various PHP errors, most notably a "passing null to non-nullable" deprecation notice on PHP 8.1.

This commit adds input validation to the `uncompress()` method, allowing only for strings and _stringable_ objects.
As the other `public` methods call `uncompress()` as the first point of action in their flow, this is sufficient to ensure input validation in all cases.

As this class was until now only indirectly tested via the `IriTest` class, a new `Ipv6Test` class is being introduced containing - for now - only perfunctory tests for the input validation.
This test class should be expanded to cover the actual logic in the methods at a later date.

**Open questions:**
* It is up for discussion whether _stringable objects_ should even be allowed as input for these methods. Opinions welcome.