Class Iri: add input validation
The documented accepted parameter type for the constructor of the `Iri` class is `string|null`, but no input validation was done on the parameter, which could lead to various PHP errors.

This commit adds input validation to the class constructor, allowing only for strings or null.

As for the other methods:
* The `public` magic methods should not need input validation as they should not be called directly, but only indirectly and when called that way, will receive the correct input type.
* The `public static` `Iri::absolutize()` method takes two parameters, both of type `\WpOrg\Requests\Iri|string` and already validates these parameters to be an instance of `Iri` and if not, invokes the class constructor, which will now automatically validate the input as `string|null`.
* The other `public` non-static methods do not take parameters.

Includes adding perfunctory tests for the new exception.