Cookie::parse(): add input validation
This adds input validation to the `Cookie::parse()` method for the `$cookie_header` and `$name` parameters.

As the `$reference_time` parameter is not directly used in this method, but only passed through to the `Cookie::__construct()` method, no input validation has been added, as it will be validated by the `Cookie::__construct()` method.

Includes tests for the new exceptions + a test to safeguard that the "fall through validation for the `$reference_time`" actually happens.

Includes fixing up two tests which were passing the wrong value type for an optional parameter.