Cookie::parse_from_headers(): add test safeguarding input validation
The first and second argument for the `Cookie::parse_from_headers()` method already have (class-based) type declarations.
The third argument - `$time` - is only used to pass it through to the `parse()` method, which will subsequently pass it through to the class constructor which contains input validation for the parameter.

This just adds a test to safeguard that the "fall through" validation for the `$time` parameter actually happens.