Cookie::path_matches(): add input validation
As this method only returns `boolean`, we may as well return boolean `false` instead of throwing an exception.

The input validation in this case is not done at the very start of the function to prevent changing the behaviour of the function for those cases which were already being handled by the method. This is also the reason for using `is_scalar()` instead of `is_string()`.

Includes tests.