Ssl: safeguard support for Stringable objects
Follow up on 572

Minor tweaks to make sure the docblocks also mention the option to pass a Stringable object and that the `Ssl::match()` method actually supports Stringable objects.

Includes adjusting a few existing tests to pass the input as Stringable objects instead of as plain strings to safeguard full Stringable support.