Session::merge_request(): improve a condition
The default value for the `$merge_options` parameter is `true`, but the condition in which the parameter is used was doing a strict comparison against `false`, which could lead to undesirable effects if `$merge_options` was passed incorrectly.

By changing the condition to do a strict comparison against the `true` default value, this issue is fixed and no additional input validation needs to be added for the `$merge_options` parameter.

Notes:
As for the `$request` parameter, I'd recommend that - at a later point in time -, stricter checks will be put in place in the `merge_requests()` method to validate that the sub-keys which are expected are of a usable type (and value) as this method is basically used as a form of input validation for the input received in the `Session::request()` and `Session::request_multiple()` methods.