Requests::request(): add input validation
This adds input validation to the `Requests::request()` method and all direct "fall-through" methods calling this method.

* The `$url` parameter is validated to allow for a string or Stringable (Iri) object.
* The `$type` parameter is validated to allow solely for a string as it expects one of the `Requests` class constants.
* The `$options` parameter is validated to allow only an array as the parameter is used with `array_merge()`.

The `$headers` and `$data` parameters remain unvalidated as they are not actually used within the method itself, but only passed through and should be validated in the `Transport::request()` method instead.

Includes adding perfunctory tests for the new exceptions + adjusting one existing test to ensure that passing the `$url` as a stringable object is safeguarded.