Request::set_certificate_path(): add input validation
Based on the documentation for the `Requests::request()` method, the `$options['verify']` allows for a string path or a boolean value, so let's allow the same for the `Requests:;set_certificate_path()` method (and include Stringable as well).

Includes perfunctory tests for the exception + for actually setting the path with valid input.

Includes:
* Removing the `DEFAULT_CERT_PATH` constant (not a BC-break as the constant was only introduced in 2.0.0-dev) and adding the default value to the `$certificate_path`.
* Simplifying the `get_certificate_path()` method.
* And using the `Requests::$certificate_path` property directly in the `get_default_options()` method as `get_certificate_path()` now no longer contains logic.